# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Restaurant Orders",
    "description" : "Simple API to manage restaurant orders",
    "contact" : {
      "name" : "Sergio Adonis",
      "email" : "sergio.adonis@outlook.com"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/sergioadonis/restaurant-orders-api/v1",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
...
    "/products" : {
      "get" : {
        "tags" : [ "Product" ],
        "operationId" : "getProducts",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 50
          }
        }, {
          "name" : "skip",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "default" : 0
          }
        }, {
          "name" : "restaurant",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        }, {
          "name" : "menu",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductArrayResult"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Product" ],
        "operationId" : "addProduct",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/productBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductObjectResult"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}" : {
      "get" : {
        "tags" : [ "Product" ],
        "operationId" : "getProductById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductObjectResult"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Product" ],
        "operationId" : "updateProduct",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "requestBody" : {
          "$ref" : "#/components/requestBodies/productBody"
        },
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductObjectResult"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Product" ],
        "operationId" : "deleteProduct",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/CommonProps/properties/id"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductObjectResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Product" : {
        "type" : "object",
        "properties" : {
          "restaurantId" : {
            "$ref" : "#/components/schemas/ReferenceId"
          },
          "price" : {
            "type" : "number",
            "format" : "currency"
          },
          "menuId" : {
            "$ref" : "#/components/schemas/ReferenceId"
          },
          "tags" : {
            "$ref" : "#/components/schemas/Tags"
          }
        },
        "example" : {
          "price" : 0.8008281904610115,
          "restaurantId" : 123,
          "tags" : [ "Mexican", "Japan", "Favorite" ]
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/CommonProps"
        } ]
      },
      "CommonProps" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64",
            "readOnly" : true,
            "example" : 123
          },
          "name" : {
            "type" : "string",
            "example" : "Name of object"
          },
          "description" : {
            "type" : "string",
            "example" : "This is a description for this object, could be Markdown sintax."
          },
          "photoUrl" : {
            "type" : "string",
            "example" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9"
          }
        },
        "example" : {
          "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
          "name" : "Name of object",
          "description" : "This is a description for this object, could be Markdown sintax.",
          "id" : 123
        }
      },
      "ReferenceId" : {
        "type" : "integer",
        "format" : "int64",
        "writeOnly" : true,
        "example" : 123
      },
      "Tags" : {
        "type" : "array",
        "example" : [ "Mexican", "Japan", "Favorite" ],
        "items" : {
          "type" : "string"
        }
      },
...
      "Result" : {
        "type" : "object",
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        }
      },
...
      "ProductArrayResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ProductArrayResult_data"
          }
        },
        "example" : {
          "data" : {
            "products" : [ {
              "price" : 0.8008281904610115,
              "restaurantId" : 123,
              "tags" : [ "Mexican", "Japan", "Favorite" ]
            }, {
              "price" : 0.8008281904610115,
              "restaurantId" : 123,
              "tags" : [ "Mexican", "Japan", "Favorite" ]
            } ]
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
      "ProductObjectResult" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ProductObjectResult_data"
          }
        },
        "example" : {
          "data" : {
            "product" : {
              "restaurant" : {
                "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
                "name" : "Name of object",
                "description" : "This is a description for this object, could be Markdown sintax.",
                "id" : 123
              }
            }
          }
        },
        "allOf" : [ {
          "$ref" : "#/components/schemas/Result"
        } ]
      },
...
      "ProductArrayResult_data" : {
        "type" : "object",
        "properties" : {
          "products" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Product"
            }
          }
        },
        "example" : {
          "products" : [ {
            "price" : 0.8008281904610115,
            "restaurantId" : 123,
            "tags" : [ "Mexican", "Japan", "Favorite" ]
          }, {
            "price" : 0.8008281904610115,
            "restaurantId" : 123,
            "tags" : [ "Mexican", "Japan", "Favorite" ]
          } ]
        }
      },
      "ProductObjectResult_data_product" : {
        "type" : "object",
        "properties" : {
          "restaurant" : {
            "$ref" : "#/components/schemas/CommonProps"
          },
          "menu" : {
            "$ref" : "#/components/schemas/CommonProps"
          }
        },
        "example" : {
          "restaurant" : {
            "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
            "name" : "Name of object",
            "description" : "This is a description for this object, could be Markdown sintax.",
            "id" : 123
          }
        }
      },
      "ProductObjectResult_data" : {
        "type" : "object",
        "properties" : {
          "product" : {
            "$ref" : "#/components/schemas/ProductObjectResult_data_product"
          }
        },
        "example" : {
          "product" : {
            "restaurant" : {
              "photoUrl" : "https://user-contents.domain.example.com/12ea34-651d76c-87bd-85b6f9",
              "name" : "Name of object",
              "description" : "This is a description for this object, could be Markdown sintax.",
              "id" : 123
            }
          }
        }
      }
    },
...
    "requestBodies" : {
...
      "productBody" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Product"
            }
          }
        },
        "required" : true
      }
    }
  }
}